Update folder tree
