+++ api/versions/v1/routers/auth/models.py
@@ -1,0 +1,12 @@
+from datetime import datetime
+from pydantic import BaseModel
+
+
+class CallbackResponse(BaseModel):
+    token: str
+    exp: datetime
+
+
+class CallbackBody(BaseModel):
+    code: str
+    callback: str
