+++ api/versions/v1/routers/router.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter
-from .auth import router as auth_router
+from .auth.routes import router as auth_router
 
 router = APIRouter(prefix="/v1")
